+++ development/src/GXDLMS.cpp
@@ -2936,7 +2936,7 @@ int CGXDLMS::HandleGloDedResponse(
             {
                 return ret;
             }
-            data.SetCipherIndex((unsigned short)data.GetData().GetSize());
+            data.SetCipherIndex(data.GetData().GetSize());
         }
     }
     return 0;
@@ -3205,7 +3205,7 @@ int CGXDLMS::GetPdu(
         {
             data.GetData().SetPosition(data.GetCipherIndex() + 1);
             ret = HandleGbt(settings, data);
-            data.SetCipherIndex((unsigned short)data.GetData().GetSize());
+            data.SetCipherIndex(data.GetData().GetSize());
             data.SetCommand(DLMS_COMMAND_NONE);
         }
         // Get command if operating as a server.
